+++ src/ds/work_beacon.h
@@ -55,6 +55,24 @@ namespace ccf::ds
 
       condition_variable.notify_all();
     }
+
+    void notify_work_available_coalesced()
+    {
+      bool notify = false;
+      {
+        ccf::pal::MutexGuard lock(mutex);
+        if (work_available == 0)
+        {
+          work_available = 1;
+          notify = true;
+        }
+      }
+
+      if (notify)
+      {
+        condition_variable.notify_all();
+      }
+    }
   };
 
   using WorkBeaconPtr = std::shared_ptr<WorkBeacon>;